Make wasi-experimental-io compile with special feature flag
The compilation of the actual code linking libxcbcommon-0 is
now gated behind --feature="link_external_libs", while the crate
continues to compile without the feature
